Added diffrent classes for error types
