+++ js-miniapp-bridge/src/index.ts
@@ -19,6 +19,12 @@ import { ScreenOrientation } from './types/screen';
 import { AccessTokenData } from './types/token-data';
 import { Contact } from './types/contact';
 import { MessageToContact } from './types/message-to-contact';
+import {
+  MiniAppError,
+  AudienceNotSupportedError,
+  AuthorizationFailureError,
+  ScopesNotSupportedError,
+} from './types/error-types';
 
 export {
   MiniAppBridge,
@@ -34,4 +40,8 @@ export {
   AccessTokenData,
   Contact,
   MessageToContact,
+  MiniAppError,
+  AudienceNotSupportedError,
+  AuthorizationFailureError,
+  ScopesNotSupportedError,
 };
