+++ js-miniapp-sdk/src/index.ts
@@ -15,15 +15,13 @@ import {
   AccessTokenData,
   Contact,
   MessageToContact,
+  MiniAppError,
+  AudienceNotSupportedError,
+  AuthorizationFailureError,
+  ScopesNotSupportedError,
 } from '../../js-miniapp-bridge/src';
 
 import { MiniApp } from './miniapp';
-import {
-  MiniAppError,
-  AuthorizationFailureError,
-  ScopesNotSupportedError,
-  AudienceNotSupportedError,
-} from './miniapp-error-type';
 
 /** @internal */
 const miniAppInstance = new MiniApp();
